flowers_bot/requirements.txt
aiogram==3.x
fastapi
uvicorn
python-dotenv
httpx
